+++ src/components/__tests__/CustomListSearch-test.tsx
@@ -1,8 +1,9 @@
 import { expect } from "chai";
-import { spy, stub } from "sinon";
+import { stub } from "sinon";
 import * as React from "react";
 import { mount } from "enzyme";
 import CustomListSearch from "../CustomListSearch";
+import CustomListSearchFormContent from "../CustomListSearchFormContent";
 
 describe("CustomListSearch", () => {
   let wrapper;
@@ -21,6 +22,8 @@ describe("CustomListSearch", () => {
     spa: ["Spanish", "Castilian"],
     fre: ["French"],
   };
+
+  const entryPoints = ["Book", "Audio"];
   beforeEach(() => {
     search = stub();
     wrapper = mount(
@@ -28,91 +31,114 @@ describe("CustomListSearch", () => {
         search={search}
         library={library}
         languages={languages}
+        entryPoints={entryPoints}
       />
     );
   });
+
   it("searches", () => {
     const input = wrapper.find(".form-control") as any;
-    input.getDOMNode().value = "test";
+    input.simulate("change", { target: { value: "test" } });
     const searchForm = wrapper.find("form");
     searchForm.simulate("submit");
 
     expect(search.callCount).to.equal(1);
-    expect(search.args[0][0]).to.equal("test");
-    expect(search.args[0][1]).to.be.null;
-  });
-  it("sorts", () => {
-    const spySort = spy(wrapper.instance(), "sort");
-    wrapper.setProps({ sort: spySort });
-    expect(wrapper.state().sortBy).to.be.null;
-    const sortOptions = wrapper.find(".search-options").find(".form-group");
-    expect(sortOptions.length).to.equal(3);
-
-    const relevance = sortOptions.at(0);
-    expect(relevance.text()).to.equal("Relevance (default)");
-    const relevanceRadio = relevance.find("input");
-    expect(relevanceRadio.props().type).to.equal("radio");
-    expect(relevanceRadio.props().name).to.be.null;
-    expect(relevanceRadio.props().value).to.equal("");
-    expect(relevanceRadio.props().checked).to.be.true;
-
-    const title = sortOptions.at(1);
-    expect(title.text()).to.equal("Title");
-    const titleRadio = title.find("input");
-    expect(titleRadio.props().type).to.equal("radio");
-    expect(titleRadio.props().name).to.equal("title");
-    expect(titleRadio.props().value).to.equal("title");
-    expect(titleRadio.props().checked).to.be.false;
-
-    const author = sortOptions.at(2);
-    expect(author.text()).to.equal("Author");
-    const authorRadio = author.find("input");
-    expect(authorRadio.props().type).to.equal("radio");
-    expect(authorRadio.props().name).to.equal("author");
-    expect(authorRadio.props().value).to.equal("author");
-    expect(authorRadio.props().checked).to.be.false;
-
-    titleRadio.simulate("change");
-    expect(spySort.callCount).to.equal(1);
-    expect(spySort.args[0][0]).to.equal("title");
-    expect(wrapper.state().sortBy).to.equal("title");
-
-    authorRadio.simulate("change");
-    expect(spySort.callCount).to.equal(2);
-    expect(spySort.args[1][0]).to.equal("author");
-    expect(wrapper.state().sortBy).to.equal("author");
-
-    relevanceRadio.simulate("change");
-    expect(spySort.callCount).to.equal(3);
-    expect(spySort.args[2][0]).to.be.null;
-    expect(wrapper.state().sortBy).to.be.null;
-
-    spySort.restore();
-  });
-  it("filters by language", () => {
-    const languageFieldset = wrapper.find("fieldset").at(2);
-    expect(languageFieldset.find("legend").text()).to.equal(
-      "Filter by language:"
+    expect(search.args[0][0]).to.equal(
+      "/short_name/search?q=test&language=all"
     );
+  });
+
+  it("searches with a language selected", () => {
+    const input = wrapper.find(".form-control") as any;
+    input.simulate("change", { target: { value: "test" } });
+
+    const languageFieldset = wrapper.find("fieldset").at(2);
     const languageMenu = languageFieldset.find("select");
-    const options = languageMenu.find("option");
-    expect(options.at(0).prop("value")).to.equal("all");
-    expect(options.at(0).text()).to.equal("All");
-    expect(options.at(1).prop("value")).to.equal("eng");
-    expect(options.at(1).text()).to.equal("English");
-    expect(options.at(2).prop("value")).to.equal("fre");
-    expect(options.at(2).text()).to.equal("French");
-    expect(options.at(3).prop("value")).to.equal("spa");
-    expect(options.at(3).text()).to.equal("Spanish; Castilian");
 
     languageMenu.getDOMNode().value = "fre";
     languageMenu.simulate("change");
-    languageFieldset.closest("form").simulate("submit");
+    const searchForm = wrapper.find("form");
+    searchForm.simulate("submit");
     expect(search.callCount).to.equal(1);
-    expect(search.args[0][2]).to.equal("fre");
+    expect(search.args[0][0]).to.equal(
+      "/short_name/search?q=test&language=fre"
+    );
   });
+
+  it("searches with ebooks selected as an entrypoint", () => {
+    const input = wrapper.find(".form-control") as any;
+    input.simulate("change", { target: { value: "test" } });
+    const radioInput = wrapper.find(".entry-points-selection input") as any;
+    const bookInput = radioInput.at(1);
+    bookInput.simulate("change");
+    const searchForm = wrapper.find("form");
+    searchForm.simulate("submit");
+    expect(search.callCount).to.equal(1);
+    expect(search.args[0][0]).to.equal(
+      "/short_name/search?q=test&entrypoint=Book&language=all"
+    );
+  });
+
+  it("searches with audiobooks selected as an entrypoint", () => {
+    const input = wrapper.find(".form-control") as any;
+    input.simulate("change", { target: { value: "test" } });
+    const radioInput = wrapper.find(".entry-points-selection input") as any;
+    const bookInput = radioInput.at(2);
+    bookInput.simulate("change");
+    const searchForm = wrapper.find("form");
+    searchForm.simulate("submit");
+    expect(search.callCount).to.equal(1);
+    expect(search.args[0][0]).to.equal(
+      "/short_name/search?q=test&entrypoint=Audio&language=all"
+    );
+  });
+
+  it("sorts", () => {
+    let formContent = wrapper.find(CustomListSearchFormContent);
+    expect(formContent.props().sortBy).to.equal(null);
+    const sortOptions = wrapper.find(".search-options").find(".form-group");
+
+    const relevance = sortOptions.at(0);
+    const relevanceRadio = relevance.find("input");
+
+    const title = sortOptions.at(1);
+    const titleRadio = title.find("input");
+
+    const author = sortOptions.at(2);
+    const authorRadio = author.find("input");
+
+    const input = wrapper.find(".form-control") as any;
+    input.simulate("change", { target: { value: "test" } });
+
+    titleRadio.simulate("change");
+    formContent = wrapper.find(CustomListSearchFormContent);
+    expect(formContent.props().sortBy).to.equal("title");
+    let searchForm = wrapper.find("form");
+    searchForm.simulate("submit");
+    expect(search.args[0][0]).to.equal(
+      "/short_name/search?q=test&order=title&language=all"
+    );
+
+    authorRadio.simulate("change");
+    formContent = wrapper.find(CustomListSearchFormContent);
+    expect(formContent.props().sortBy).to.equal("author");
+    searchForm = wrapper.find("form");
+    searchForm.simulate("submit");
+    expect(search.args[1][0]).to.equal(
+      "/short_name/search?q=test&order=author&language=all"
+    );
+
+    relevanceRadio.simulate("change");
+    formContent = wrapper.find(CustomListSearchFormContent);
+    expect(formContent.props().sortBy).to.equal(null);
+    searchForm = wrapper.find("form");
+    searchForm.simulate("submit");
+    expect(search.args[2][0]).to.equal(
+      "/short_name/search?q=test&language=all"
+    );
+  });
+
   it("automatically searches if there is a startingTitle prop", () => {
-    const search = stub();
     wrapper = mount(
       <CustomListSearch
         startingTitle="test"
@@ -122,7 +148,8 @@ describe("CustomListSearch", () => {
       />
     );
     expect(search.callCount).to.equal(1);
-    expect(search.args[0][0]).to.equal("test");
-    expect(search.args[0][1]).to.be.null;
+    expect(search.args[0][0]).to.equal(
+      "/short_name/search?q=test&language=all"
+    );
   });
 });