+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ## Changelog
 
-### 11/23/2021
+### v0.5.10
 
 #### Refactored
 
+- Refactored the `CustomListSearch` component to bring it up-to-date with current best practices.
+- Added `CustomListSearchFormContent` and `CustomListSearchAdvancedOptions` to make components smaller.
+- Added/edited tests where needed.
 - Created reusable component `CustomListBookCard` to reduce repetitive code in `CustomListSearchResults` and `CustomListEntries`.
 
 ### v0.5.9